Update stat_generator to generate age and race
This is an update to generate more of the character traits. 
This will select a race in addition to stats and age. Since age is influenced by race, it must be chosen after. 
A "Life Stage" has been introduced to contextualize the character's age in regard to it's race. 

The file also writes to a csv now.